backend/requirements.txt
pyyaml>=6.0
requests>=2.31
google-api-python-client>=2.100
google-auth>=2.23
google-auth-oauthlib>=1.1
google-auth-httplib2>=0.2
gspread>=6.0
Pillow>=10.0
